projeto-1-script/requirements.txt
web3==6.20.1    
python-dotenv==1.0.1